Uploading the csv files for 1, 3, and 5 KPCA
Each csv was used to test and train the ML/DL algorithms with 1, 3, and 5 KPCA extracted features from the images. The best model was found to be LH_KPCA_5_highRes.csv file trained on split MLP model. If one runs remove_background.py and KPCA.ipynb they will get data that is similar to this, as the KPCA feature reduction will not be the same every time as some data might be lost when compressing.