Uploading notebooks for models.
Imporving.ipynb, might need name change, but contains Feature Importance analysis on random forest, and contains random forest and XGBoost models using LH data to predict the drag coefficient. LH_MLP.ipynb is a notebook that contains PyTorch code for creating a neural network for the prediction of drag coefficient with custom pytorch dataset and model. Split_MLP_LH.ipynb is a notebook that contains PyTorch code for the split MLP model on Reynolds Number for the LH dataset.